Rehauled part of a package structure
