+++ app/src/main/java/com/vultisig/wallet/app/activity/MainViewModel.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.vultisig.wallet.data.on_board.db.VaultDB
-import com.vultisig.wallet.on_board.OnBoardRepository
+import com.vultisig.wallet.data.repositories.OnBoardRepository
 import com.vultisig.wallet.ui.navigation.Destination
 import com.vultisig.wallet.ui.navigation.Navigator
 import com.vultisig.wallet.ui.navigation.Screen
