--- app/src/main/java/com/vultisig/wallet/data/on_board/mappers/OnBoardPage.kt
@@ -1,5 +1,0 @@
-package com.vultisig.wallet.data.on_board.mappers
-
-import com.vultisig.wallet.data.on_board.models.OnBoardPage
-
-fun OnBoardPage.toDomain(): OnBoardPage = OnBoardPage(image, title, description)
