+++ app/src/main/java/com/vultisig/wallet/data/repositories/RepositoriesModule.kt
@@ -44,6 +44,12 @@ internal interface RepositoriesModule {
         impl: BalanceRepositoryImpl,
     ): BalanceRepository
 
+    @Singleton
+    @Binds
+    abstract fun bindOnBoardRepository(
+        impl: DataStoreRepositoryImpl,
+    ): OnBoardRepository
+
     companion object {
 
         @Provides
