+++ app/src/main/java/com/vultisig/wallet/presenter/keygen/DeviceList.kt
@@ -15,9 +15,9 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import com.vultisig.wallet.R
 import com.vultisig.wallet.common.Utils
-import com.vultisig.wallet.presenter.base_components.MultiColorButton
-import com.vultisig.wallet.presenter.common.TopBar
 import com.vultisig.wallet.presenter.keygen.components.DeviceInfoItem
+import com.vultisig.wallet.ui.components.MultiColorButton
+import com.vultisig.wallet.ui.components.TopBar
 import com.vultisig.wallet.ui.theme.appColor
 import com.vultisig.wallet.ui.theme.dimens
 import com.vultisig.wallet.ui.theme.menloFamily
