+++ app/src/main/java/com/vultisig/wallet/presenter/welcome/WelcomeScreen.kt
@@ -30,11 +30,11 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.vultisig.wallet.R
-import com.vultisig.wallet.data.on_board.models.OnBoardPage
-import com.vultisig.wallet.presenter.base_components.MultiColorButton
+import com.vultisig.wallet.data.models.OnBoardPage
 import com.vultisig.wallet.presenter.common.UiEvent.NavigateTo
 import com.vultisig.wallet.presenter.common.UiEvent.PopBackStack
 import com.vultisig.wallet.presenter.common.UiEvent.ScrollToNextPage
+import com.vultisig.wallet.ui.components.MultiColorButton
 import com.vultisig.wallet.ui.theme.appColor
 import com.vultisig.wallet.ui.theme.dimens
 import com.vultisig.wallet.ui.theme.montserratFamily
