+++ app/src/main/java/com/vultisig/wallet/presenter/welcome/WelcomeState.kt
@@ -1,6 +1,6 @@
 package com.vultisig.wallet.presenter.welcome
 
-import com.vultisig.wallet.data.on_board.models.OnBoardPage
+import com.vultisig.wallet.data.models.OnBoardPage
 
 data class WelcomeState(
     val pages: List<OnBoardPage> = emptyList(),
