+++ app/src/main/java/com/vultisig/wallet/ui/components/multi_color_button.kt
@@ -1,4 +1,4 @@
-package com.vultisig.wallet.presenter.base_components
+package com.vultisig.wallet.ui.components
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
