+++ app/src/main/java/com/vultisig/wallet/ui/screens/ChainCoinScreen.kt
@@ -1,4 +1,4 @@
-package com.vultisig.wallet.presenter.chain_coin
+package com.vultisig.wallet.ui.screens
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
@@ -50,8 +50,9 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.vultisig.wallet.R
-import com.vultisig.wallet.presenter.base_components.MultiColorButton
+import com.vultisig.wallet.ui.components.MultiColorButton
 import com.vultisig.wallet.ui.components.UiPlusButton
+import com.vultisig.wallet.ui.models.ChainCoinViewModel
 import com.vultisig.wallet.ui.navigation.Screen
 import com.vultisig.wallet.ui.theme.Theme
 import com.vultisig.wallet.ui.theme.appColor
