+++ app/src/main/java/com/vultisig/wallet/ui/screens/ChainSelectionScreen.kt
@@ -19,8 +19,8 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.vultisig.wallet.R
 import com.vultisig.wallet.models.logo
-import com.vultisig.wallet.presenter.common.TopBar
 import com.vultisig.wallet.ui.components.TokenSelectionItem
+import com.vultisig.wallet.ui.components.TopBar
 import com.vultisig.wallet.ui.models.ChainSelectionViewModel
 import com.vultisig.wallet.ui.theme.appColor
