+++ app/src/main/java/com/vultisig/wallet/ui/screens/VaultSettingsScreen.kt
@@ -14,8 +14,8 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import com.vultisig.wallet.R
-import com.vultisig.wallet.presenter.common.TopBar
 import com.vultisig.wallet.ui.components.SettingsItem
+import com.vultisig.wallet.ui.components.TopBar
 import com.vultisig.wallet.ui.theme.appColor
 
 @Composable
